Use all aggregators for parallel read
Summary:
For better load balancing we need to connect to different aggregators during parallel read. In this diff we introduce a mechanism to do so. Partition `p` is assigned to the aggregator number `p % (number_of_aggregators)`, and each `ResultTableReader` gets its own credentials to connect to the server.
With this change version 1.0.0 is released.

Test Plan: ./build_and_test.sh test

Reviewers: amakarovych-ua

Reviewed By: amakarovych-ua

Subscribers: eli, engineering-list

JIRA Issues: PLAT-5434

Differential Revision: https://grizzly.internal.memcompute.com/D53706